Moving functions around to expose to other web services
