+++ src/Nullinside.Api.Common/Nullinside.Api.Common.csproj
@@ -15,4 +15,11 @@
         <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
     </PropertyGroup>
 
+    <ItemGroup>
+        <PackageReference Include="Microsoft.Extensions.Logging" Version="8.0.0"/>
+        <PackageReference Include="Newtonsoft.Json" Version="13.0.3"/>
+        <PackageReference Include="SSH.NET" Version="2024.0.0"/>
+        <PackageReference Include="TwitchLib.Api" Version="3.9.0"/>
+    </ItemGroup>
+
 </Project>
